pyproject.toml
[tool.poetry]
name = "cpython-stats"
version = "21.8.0"
description = "What are all those people doing?"
authors = ["Łukasz Langa <lukasz@langa.pl>"]

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.55"
python-dotenv = "^0.19.0"
rich = "^10.7.0"
python-dateutil = "^2.8.2"
sqlite-utils = "^3.17"
tomli = "^1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.7b0"
mypy = "^0.910"
flake8-bugbear = "^21.4.3"
types-python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
